Run wheel test only when multi-GPU flag is not set

+++ Jenkinsfile
@@ -85,14 +85,18 @@ def buildPlatformCmake(buildName, conf, nodeReq, dockerTarget) {
         sh """
         ${dockerRun} ${dockerTarget} ${dockerArgs} tests/ci_build/build_via_cmake.sh ${opts}
         ${dockerRun} ${dockerTarget} ${dockerArgs} tests/ci_build/test_${test_suite}.sh
-        ${dockerRun} ${dockerTarget} ${dockerArgs} bash -c "cd python-package; rm -f dist/*; python setup.py bdist_wheel --universal"
-        rm -rf "${distDir}"; mkdir -p "${distDir}/py"
-        cp xgboost "${distDir}"
-        cp -r python-package/dist "${distDir}/py"
-        # Test the wheel for compatibility on a barebones CPU container
-        ${dockerRun} release ${dockerArgs} bash -c " \
-            pip install --user python-package/dist/xgboost-*-none-any.whl && \
-            python -m nose tests/python"
         """
+        if (!conf["multiGpu"]) {
+            sh """
+            ${dockerRun} ${dockerTarget} ${dockerArgs} bash -c "cd python-package; rm -f dist/*; python setup.py bdist_wheel --universal"
+            rm -rf "${distDir}"; mkdir -p "${distDir}/py"
+            cp xgboost "${distDir}"
+            cp -r python-package/dist "${distDir}/py"
+            # Test the wheel for compatibility on a barebones CPU container
+            ${dockerRun} release ${dockerArgs} bash -c " \
+                pip install --user python-package/dist/xgboost-*-none-any.whl && \
+                python -m nose tests/python"
+            """
+        }
     }
 }
